CRITICAL: Move all other models to Queue DB.

See merge request crafty-controller/crafty-commander!241

+++ app/classes/models/server_permissions.py
@@ -8,7 +8,6 @@ from app.classes.shared.permission_helper import permission_helper
 
 try:
     from peewee import (
-        SqliteDatabase,
         Model,
         ForeignKeyField,
         CharField,
@@ -16,6 +15,7 @@ try:
         JOIN,
     )
     from enum import Enum
+    from playhouse.sqliteq import SqliteQueueDatabase
 
 except ModuleNotFoundError as e:
     helper.auto_installer_fix(e)
@@ -23,8 +23,9 @@ except ModuleNotFoundError as e:
 logger = logging.getLogger(__name__)
 peewee_logger = logging.getLogger("peewee")
 peewee_logger.setLevel(logging.INFO)
-database = SqliteDatabase(
-    helper.db_path, pragmas={"journal_mode": "wal", "cache_size": -1024 * 10}
+database = SqliteQueueDatabase(
+    helper.db_path
+    # pragmas={"journal_mode": "wal", "cache_size": -1024 * 10}
 )
 
 # **********************************************************************************
@@ -193,12 +194,9 @@ class Permissions_Servers:
 
     @staticmethod
     def remove_roles_of_server(server_id):
-        with database.atomic():
-            return (
-                Role_Servers.delete()
-                .where(Role_Servers.server_id == server_id)
-                .execute()
-            )
+        return (
+            Role_Servers.delete().where(Role_Servers.server_id == server_id).execute()
+        )
 
     @staticmethod
     def get_user_id_permissions_mask(user_id, server_id: str):
